requirements.txt
mysql-connector-python
pandas
matplotlib
scikit-learn
